# openapi.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "ErrorResponse": {
      "properties": {
        "error": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
    "RouteCompareResponse": {
      "properties": {
        "comparison": {
          "properties": {
            "best_route_index": {
              "type": "integer"
            },
            "best_score": {
              "type": "number"
            },
            "route_count": {
              "type": "integer"
            },
            "score_range": {
              "properties": {
                "average": {
                  "type": "number"
                },
                "max": {
                  "type": "number"
                },
                "min": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "scores": {
              "items": {
                "$ref": "#/definitions/RouteScoreData"
              },
              "type": "array"
            }
          },
          "type": "object"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
    "RouteItem": {
      "$ref": "#/definitions/Store"
    },
    "RouteScoreData": {
      "properties": {
        "calculation_time_ms": {
          "type": "number"
        },
        "component_scores": {
          "additionalProperties": {
            "type": "number"
          },
          "type": "object"
        },
        "metadata": {
          "type": "object"
        },
        "raw_metrics": {
          "type": "object"
        },
        "total_score": {
          "type": "number"
        }
      },
      "type": "object"
    },
    "RouteScoreResponse": {
      "properties": {
        "score": {
          "$ref": "#/definitions/RouteScoreData"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
    "Store": {
      "properties": {
        "id": {
          "type": "string"
        },
        "lat": {
          "type": "number"
        },
        "lon": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "priority": {
          "type": "integer"
        }
      },
      "required": [
        "lat",
        "lon"
      ],
      "type": "object"
    },
...
  },
  "info": {
    "contact": {
      "name": "RouteForce Support",
      "url": "https://github.com/ApacheEcho/RouteForceRouting"
    },
    "description": "API documentation for RouteForce Routing application - A comprehensive route optimization platform for field execution teams",
    "title": "RouteForce Routing API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/mobile/routes/{route_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Route details returned"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Route not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get detailed route information",
        "tags": [
          "Mobile"
        ]
      }
    },
...
    "/api/route/score": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "context": {
                  "type": "object"
                },
                "preset": {
                  "enum": [
                    "balanced",
                    "distance_focused",
                    "priority_focused",
                    "traffic_aware",
                    "playbook_strict"
                  ],
                  "type": "string"
                },
                "route": {
                  "items": {
                    "$ref": "#/definitions/RouteItem"
                  },
                  "type": "array"
                },
                "weights": {
                  "type": "object"
                }
              },
              "required": [
                "route"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Score computed",
            "examples": {
              "application/json": {
                "score": {
                  "calculation_time_ms": 4.12,
                  "component_scores": {
                    "distance": 0.92,
                    "efficiency": 0.7,
                    "playbook": 0.8,
                    "priority": 0.9,
                    "time": 0.85,
                    "traffic": 0.78
                  },
                  "metadata": {
                    "route_length": 5
                  },
                  "raw_metrics": {
                    "total_distance_km": 12.3,
                    "total_time_h": 0.75
                  },
                  "total_score": 87.25
                },
                "success": true
              }
            },
            "schema": {
              "$ref": "#/definitions/RouteScoreResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Score a single route",
        "tags": [
          "Routes"
        ]
      }
    },
    "/api/route/score/compare": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "context": {
                  "type": "object"
                },
                "preset": {
                  "type": "string"
                },
                "routes": {
                  "items": {
                    "items": {
                      "$ref": "#/definitions/RouteItem"
                    },
                    "type": "array"
                  },
                  "type": "array"
                }
              },
              "required": [
                "routes"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comparison computed",
            "examples": {
              "application/json": {
                "comparison": {
                  "best_route_index": 0,
                  "best_score": 82.5,
                  "route_count": 2,
                  "score_range": {
                    "average": 80.3,
                    "max": 82.5,
                    "min": 78.1
                  },
                  "scores": [
                    {
                      "calculation_time_ms": 3.2,
                      "component_scores": {
                        "distance": 0.9
                      },
                      "metadata": {},
                      "raw_metrics": {},
                      "total_score": 82.5
                    },
                    {
                      "calculation_time_ms": 3.4,
                      "component_scores": {
                        "distance": 0.85
                      },
                      "metadata": {},
                      "raw_metrics": {},
                      "total_score": 78.1
                    }
                  ]
                },
                "success": true
              }
            },
            "schema": {
              "$ref": "#/definitions/RouteCompareResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Compare multiple routes",
        "tags": [
          "Routes"
        ]
      }
    },
...
    "/api/v1/routes": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "constraints": {
                  "type": "object"
                },
                "options": {
                  "properties": {
                    "algorithm": {
                      "enum": [
                        "default",
                        "nearest_neighbor",
                        "priority",
                        "genetic",
                        "simulated_annealing",
                        "multi_objective"
                      ],
                      "type": "string"
                    },
                    "ga_generations": {
                      "type": "integer"
                    },
                    "ga_population_size": {
                      "type": "integer"
                    },
                    "mo_objectives": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                },
                "stores": {
                  "items": {
                    "$ref": "#/definitions/Store"
                  },
                  "type": "array"
                }
              },
              "required": [
                "stores"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Route created successfully",
            "examples": {
              "application/json": {
                "data": {
                  "algorithm_used": "genetic",
                  "route": [
                    {
                      "id": "s1",
                      "lat": 40.71,
                      "lon": -74.0,
                      "name": "Store A"
                    },
                    {
                      "id": "s2",
                      "lat": 40.76,
                      "lon": -73.98,
                      "name": "Store B"
                    }
                  ],
                  "route_id": 101
                },
                "metadata": {
                  "optimization_score": 1.8,
                  "processing_time": 0.25,
                  "route_stores": 2,
                  "total_stores": 2
                },
                "success": true
              }
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Create a new route via API with database persistence",
        "tags": [
          "Routes"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "in": "header",
      "name": "X-API-Key",
      "type": "apiKey"
    },
    "JWT": {
      "description": "JWT Bearer token. Use: 'Bearer <token>'",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}